requirements.txt
streamlit
google-generativeai
PyPDF2
python-docx
plotly
pandas
python-dotenv
